feat(owner): refactor AccommodationInfoField and RoomInfoField to use CommonFields and StepperField components
